W462: branch-polarity census; nd -1,007, 22 functions improved, 0 regressed
New instrument: tools/branch_census.py finds branches retail tests the opposite
way round from us -- beq<->bne, beql<->bnel, blez<->bgtz, bltz<->bgez,
bc1f<->bc1t -- after difflib alignment.  Inverting a source condition and
swapping its two arms is exact and legal; it is not the banned change of a
comparison's *meaning*, and the test is whether any input yields a different
outcome.

Catalogue-wide 450 functions show a polarity difference but only **32 have a
FLIPPED row before any insertion**.  Those 32 were the actionable set.

Gate (build/baselines/W462_gate.json vs W461_gate.json):

  first-party MATCH      6,930   (unchanged)
  first-party nd         1,191,181 -> 1,190,174   (-1,007)
  MATCH lost             0
  functions improved     22
  functions regressed    0
  over-window            3 -> 3
  nd-0 but not MATCH     0
  pytest                 92 passed
  build                  SLUS_216.21 sha1 3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb  OK
  reloc_audit            0 wrong call targets

**Measured hit rate for the mechanism: eight firm flips in one lane gave four
positives and four negatives.**  Across all lanes it is roughly half.  That
number matters more than any single result -- it stops the next wave treating a
flagged flip as a guaranteed win.

Largest gains:

  k_vpad      func_001ebd80  2738 -> 2580
  y_misc      FUN_004534b0   1625 -> 1476   u32 -> u8 channel local
  k_unit      func_001d0720   160 ->   49
  itfMes      FUN_003a5980    202 ->  118   three successive mask inversions,
                                            object 800/800 throughout
  fclComb     FUN_003d6c90    209 ->  132   inversion, then ascending
                                            switch(param_1) 0/1/2, then
                                            int->u32, then explicit int casts
                                            aligning sltiu -> slti
  fclComb     FUN_003d72f0    246 ->  179
  k_dungeon   func_001bcac0  1551 -> 1492
  frTagTable  FUN_003af110    358 ->  300
  btlUnit     btlUnitUpdateModelPacket  892 -> 846
  y_timeLimit FUN_0045c530    561 ->  521
  fclShopMisc FUN_00402c80/00402800  twins, 697 -> 612 / 669
  mdlEffect   FUN_00343d60   1730 -> 1705 and FUN_00329ed0 226 -> 218
  btlFormation func_002ba3a0  444 ->  421   at 1376/1376, zero margin
  mt_evtLoadSave FUN_0038ffb0 395 -> 377

Two sub-patterns worth naming:

**Early-return guard to wrapping conditional.**  `if (c) return; body;` becomes
`if (!c) { body }`, which removes one epilogue and often flips the branch to
retail's polarity.  Landed in k_dungeon (-59, object -8) and btlUnit (-46,
object -12).  Both shrinks were audited statement-by-statement: the line-count
delta accounts for braces only, so this is not the W460 k_vpad failure mode
where "merging duplicate branches" deleted 648 real bytes.

**The chain compounds.**  fclCombineMisc FUN_003d6c90 took four successive
source corrections, each exposing the next, for -77 total.  itfMesManager
FUN_003a5980 took three inversions in one function for -84 -- when *every*
condition in a function is written the opposite way round, the original
probably had a different overall shape.

Notable negatives, all with numbers in the lane reports:

  - fclShopMisc FUN_003fc540: an inverted giant arm measures 757 -> 607 but the
    object lands at 1,096 against a 1,088 window.  **8 bytes from the largest
    unclaimed gain in the catalogue.**
  - mt_evtLoadSave: the same transformation (u64 param + pre-call slot pointer)
    gains 18 nd on FUN_0038ffb0 and costs 100 on FUN_003902c0, two structurally
    near-identical functions differing in a constant and the array element type.
    Neither half of the change does anything alone.
  - k_battle: retail's `valid` boolean merge (sltu/beqz rather than short-circuit
    branches) was reconstructed from the disassembly and measured on all three
    siblings -- +2 and over window, +13 with a shrink, and flat.  The shape is
    right and the source form is not.
  - btlMain FUN_0029ee20: four float-comparison rewrites, best +1.  The `bc1f`
    is not reachable by operand or arm reordering.
  - K_Fldrc_Init: an arm swap grew the object 12 bytes and went over window,
    which means MWCC generated a different construct rather than a mirrored one.

Incident: the jj working copy went stale mid-wave (`not updated since operation
785303440097`), so `jj file show` returned empty and several restore
confirmations were wrong.  Recovered by backing up src/, running
`jj workspace update-stale`, diffing the result, and re-applying the four files
jj had reverted.  No lane work was lost.  Two lanes independently noticed the
divergence from their own readings before I did.

Lane reports: build/wip/*W462_report.md.

+++ src/Yajima/y_misc.c
@@ -28864,7 +28864,7 @@ u32 FUN_004534b0(char param_1)
 {
   u8 bVar1;
   u8 *puVar2;
-  u32 uVar3;
+  u8 uVar3;
   u32 uVar4;
   int iVar5;
   int lVar6;
